% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\docType{class}
\name{Database-class}
\alias{Database-class}
\alias{Database}
\title{Abstract base class for databases}
\description{
This class is an interface defining the functions that database classes (db_xxx.R) must implement.
It is not intented for use by end-users, however if you wish to add your own database to this
package, you can create a file db_databaseauthor_databasename.R and add a class that "contains" 
this class, in a similar manner to how the other db_xxx.R files are constructed.
}
\details{
The functions that subclasses (classes in db_xxx.R files) must implement are the functions
which contain the "stop" function, which means that an error will be thrown if the subclass
does not implement that function.
}
\section{Fields}{

\describe{
\item{\code{dir}}{character containing the path to this database's folder where it can store data.}
}}


