% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Databases.R
\docType{class}
\name{Databases-class}
\alias{Databases-class}
\alias{Databases}
\title{Search, download manage databases}
\description{
The class \code{Databases} handles initialising databases (eg. downloading relevant files)
and dispatching requests to each database for searching them.
}
\details{
Note that this class does not handle synonym-checking; this is performed by the 
\code{\link{find_species_traits}} function. This class also has a custom $new function, which
takes two arguments: \code{directory}, which is the directory that 
you want to store downloaded database files in (usually CSV or Excel files); and 
\code{database_classes}, which is a list of the databases you would like it to 
manage. It defaults to \\code{functionaltraits_db_list}, which includes all available databases.
If a database is causing errors or cannot connect to the server and is causing the ready() function
to fail, then you can omit it from this argument and it will not be included."
}
\section{Fields}{

\describe{
\item{\code{dir}}{character containing the path to a folder that all of the database files will be downloaded to}

\item{\code{databases}}{a list with the names of the databases as the key and the value being the actual
class of the database.}
}}

\section{Methods}{

\describe{
\item{\code{check_traits(traits)}}{This function takes one argument, \code{traits}, and checks whether it is a validly formatted
list. The names of the list elements must correspond to the names of the databases as provided
in \code{database_classes} to the $new function, and the elements must be vectors containing
the names of the columns selected from the database. If the list is not valid, this function
will throw an error explaining why.}

\item{\code{columns()}}{This function returns a list of all columns that are available for each database, in
a format suitable to be provided to $search or $check_traits. You can use it to find out
what columns are available and what their names are.}

\item{\code{drop(database_name)}}{This function removes a database. It can be used if a database cannot properly configure itself using the $initialise() 
function and consequently would not return TRUE when $ready() is called. It takes one argument, database_name, which is 
a string containing the name of the database that you want to remove.}

\item{\code{initialise()}}{This functions initialises the databases if they are not already initialised.}

\item{\code{ready()}}{This function checks whether all the databases are ready for use. If they are, it returns TRUE.
If not, it returns FALSE. Each database that is not ready will print a warning message.}

\item{\code{search(species, traits = NULL)}}{This function takes one argument, \code{species} and searches the databases for those species.
It can optionally take another argument, \code{traits}, which is a list of the traits to return 
from each database. It is a list of vectors, where the name of each list element is
the name of a database and the vectors hold the names of the traits to return, 
eg. list( fishbase = c( 'Weight' ) ). 

It returns a list, with one element \code{results} containing a dataframe of the results; 
and another element \code{statistics} being a dataframe with the number of matches and columns
selected from each database.}
}}

