% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trait_databases.R
\name{download_trait_databases}
\alias{download_trait_databases}
\title{Download trait databases}
\usage{
download_trait_databases()
}
\value{
No return value.
}
\description{
Downloads the trait databases used by \code{\link{find_species_traits}} into a folder
called 'data' in the current working directory.
}
\examples{
download_trait_databases()
}
