% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_species_traits.R
\name{find_species_traits}
\alias{find_species_traits}
\title{Find traits and taxonomic information for a list of species}
\usage{
find_species_traits(databases, species, traits = NULL,
  get_common_names = FALSE)
}
\arguments{
\item{species}{A vector of species names in the format "Genus species". Note: if the same name 
is given more than once, the extra occurances are discarded. Consequently, you
cannot be guaranteed that the number of rows in the resulting dataframe has the same
length as the number of species that were given.

Note that it also removes duplicate species names that are given. It only takes
species names, NOT genus or family names. If genus or other taxonomic ranks are given,
they are ignored.}

\item{traits=NULL}{A list of trait databases, with each value being a vector of column names. Eg.
\code{list( "fishbase" = c( "Weight" ) )} to search fishbase and include the
Weight column in the output. If \code{NULL}, then all databases and all columns
will be selected.}

\item{get_common_names=NULL}{Whether to get common name information as well. This significantly increases the time
that this function takes to run.}
}
\value{
A \code{list} with \code{list[["results"]]} being a dataframe containing the list of species
        and any matched trait data from the selected columns; and \code{list[["statistics"]]} being
        a dataframe containing the number of matches and number of columns selected from each database.
}
\description{
This function takes a list of species, collects taxonomic information for them, 
and then searches the databases for the valid scientific names, as well as any 
synonyms that were found.
}
\details{
The main difference between this function, and \code{Databases::search()}, is that 
it also collects taxonomic information, and searches the databases for synonyms.

It returns a dataframe with the following columns:
- taxa: the species name given in the \code{species} argument
- found: whether or not the species name was recognised by the Catalogue of Life (https://www.catalogueoflife.org/)
         taxonomic database.
- colid: the ID for the taxonomic name in the Catalogue of Life
- synonyms: any synonyms found
- accepted_name: the accepted name for the species (according to Catalogue of Life)
- common_name: the common name(s) of the species
- kingdom, phylum, class, order, family, genus: taxonomic information when available

In addition, it includes all the columns selects via the \code{traits} argument. Columns are preprended
with database author's name to ensure that the source of the information can be traced and appropriately
cited.
}
\examples{
find_species_traits( c("Betta splendens"), list( "fishbase" = c( "Weight" ) ))
find_species_traits( c("Betta splendens"), list( "pacifici_generationlength" = c( "GenerationLength_d" ) ))
find_species_traits( c("Betta splendens"))
}
