% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchResults.R
\docType{class}
\name{SearchResults-class}
\alias{SearchResults-class}
\alias{SearchResults}
\title{An S4 class to represent the search results from a single database}
\description{
An S4 class to represent the search results from a single database
}
\section{Slots}{

\describe{
\item{\code{results}}{A dataframe containing a column, \code{species}, containing the list of species that was
given to the database}

\item{\code{numberOfMathces}}{The number of species that trait information was found for in this database}

\item{\code{numberOfColumns}}{The number of columns that were selected from this database}
}}

