% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_species_traits.R
\name{find_species_traits}
\alias{find_species_traits}
\title{Find traits for a list of species}
\usage{
find_species_traits(species, attributes = NULL)
}
\arguments{
\item{species}{A vector of species names in the format "Genus species"}

\item{attributes=NULL}{A list of trait databases, with each value being a vector of column names. Eg.
\code{list( "fishbase" = c( "Weight" ) )} to search fishbase and include the
Weight column in the output. If \code{NULL}, then all databases and all columns
will be selected.}
}
\value{
A \code{list} with \code{list[["traits"]]} being a dataframe containing the list of species
        and any matched trait data from the selected columns.
}
\description{
Find traits for a list of species
}
\examples{
find_species_traits( c("Betta splendens"), list( "fishbase" = c( "Weight" ) ))
find_species_traits( c("Betta splendens"), list( "pacifici_generationlength" = c( "GenerationLength_d" ) ))
find_species_traits( c("Betta splendens"))
}
